+++ 404.php
@@ -20,7 +20,7 @@ get_header();
     <div class="main-content not-found-content">
     
     <h1>404</h1>
-    <p><?php esc_attr_e( "It seems that the page you were trying to reach couldn't be found, or maybe you have mis-typed the URL, please check your spelling. We think that the best thing to do is to use the search bar at the upper right of this page, or check out our homepage. Please contact us if the problem persists or if you definitely can't find what you are looking for. Thank you very much." ); ?></p>
+    <p><?php esc_attr_e( "It seems that the page you were trying to reach couldn't be found, or maybe you have mis-typed the URL, please check your spelling. We think that the best thing to do is to use the search bar at the upper right of this page, or check out our homepage. Please contact us if the problem persists or if you definitely can't find what you are looking for. Thank you very much.", 'geckopress' ); ?></p>
     
     </div>