+++ searchform.php
@@ -2,7 +2,7 @@
 /**
  * Template file for displaying search form.
  *
- * This search form can be displayed in the navigation menu or you can use the built-in WordPress search widget.
+ * This search form can be displayed with the built-in WordPress search widget.
  *
  * @package GeckoPress
  * @subpackage Main
@@ -15,9 +15,9 @@ defined( 'ABSPATH' ) or die( 'Cannot access pages directly.' );
 
 ?>
 <form role="search" method="get" class="search-form form" action="<?php echo esc_url( home_url( '/' ) ); ?>">
-	<label for="form-search-input" class="sr-only"><?php esc_html_ex( 'Search for', 'label', 'geckopress' ); ?></label>
+	<label for="form-search-input" class="sr-only"><?php esc_html_e( 'Search for', 'geckopress' ); ?></label>
 	<div class="input-group">
-		<input type="search" id="form-search-input" class="form-control" placeholder="<?php echo esc_attr_x( 'Search &hellip;', 'placeholder', 'bootstrap-basic' ); ?>" value="<?php echo esc_attr( get_search_query() ); ?>" name="s" title="<?php echo esc_attr_x( 'Search for:', 'label', 'geckopress' ); ?>">
+		<input type="search" id="form-search-input" class="form-control" placeholder="<?php esc_attr_e( 'Search &hellip;', 'geckopress' ); ?>" value="<?php echo esc_attr( get_search_query() ); ?>" name="s" title="<?php esc_attr_e( 'Search for:', 'geckopress' ); ?>">
 		<span class="input-group-btn">
 			<button type="submit" class="btn btn-default"><?php esc_html_e( 'Search', 'geckopress' ); ?></button>
 		</span>