+++ snippets/loops/post-head-meta.php
@@ -2,7 +2,7 @@
     <p>
         <?php 
             
-            $str = __('Posted on %s by %s');
+            $str = __('Posted on %s by %s', 'geckopress');
             $date = get_the_date(); 
             $author = get_the_author();
             echo sprintf($str, $date, $author);