+++ snippets/loops/single-page.php
@@ -1,5 +1,9 @@
+<div id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
+    
 <h1><?php the_title(); ?></h1>
 
 <div class="post-content">
     <?php the_content(); ?>
+</div>
+
 </div>