+++ functions/scripts.php
@@ -7,6 +7,12 @@ function gp_register_scripts()
         array('jquery'),
         '',
         true);
+        
+    wp_register_script('gp_base',
+        get_template_directory_uri() . '/assets/js/geckopress-base.js',
+        array('jquery'),
+        '',
+        true);
 }
 
 function gp_enqueue_scripts()
@@ -14,6 +20,11 @@ function gp_enqueue_scripts()
     if(!is_admin() && GPConfig::getInstance()->get_cfg('gp_enqueue_bootstrap'))
     {
         wp_enqueue_script('bootstrap');
+        
+        if(GPConfig::getInstance()->get_cfg('gp_enqueue_base'))
+        {
+            wp_enqueue_script('gp_base');
+        }
     }
 }