+++ snippets/loops/archive-search.php
@@ -1,4 +1,4 @@
-<div class="search-item">
+<div id="post-<?php the_ID(); ?>" <?php post_class('search-item'); ?>>
     <h2>
         <a href="<?php the_permalink(); ?>" title="<?php the_title(); ?>">
             <?php the_title(); ?>